Worldgen hang fix (BoP now with mushrooms!)
Problem was two fold:
Primary issue: var2 was being used in the for loop, but also being set
in the for loop. this was creating an infinite loop.
Secondary issue: mushroom*Gen was not being initialized. This was being
caught by the try...catch, but if the primary issue had been fixed alone
there would still be no mushrooms.